requirements.txt
flask
flask-login
mongoengine
gunicorn
python-dotenv
werkzeug
cloudinary
itsdangerous
